coral survivability seems ok for now

+++ RecoveReef Game/Assets/Resources/SubstrataData.cs
@@ -8,7 +8,7 @@ public class SubstrataData
     [XmlAttribute("name")]
     public string name;
     [XmlElement("GroundViability")]
-    public float groundViability;
+    public int groundViability;
 
     public string dataToString() {
         string output = "---SubstrataData---";
